Price BigDecimal

+++ src/main/java/com/andersen_intensive/hotel/models/IApartment.java
@@ -1,9 +1,11 @@
 package com.andersen_intensive.hotel.models;
 
 
+import java.math.BigDecimal;
+
 public interface IApartment {
     public int getApartmentNumber();
-    public Double getApartmentPrice();
+    public BigDecimal getApartmentPrice();
     public ApartmentType getApartmentType();
     public ApartmentStatus getApartmentStatus();
